Add additional guard clause to annotate_span_with_response! in net_http instrumentation
We want to make sure we can co-exist with other tracing libraries within
a Rails application.
annotate_span_with_response! would get called not only for our own spans
but also for spans coming from other Tracing libraries such as DataDog.
To make sure we don't fail in case this method gets called with a DD
span we check if the span responds to the method we need